# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  models.Contact:
    properties:
      allocated_amount:
        description: 'Computed: Sum of payments'
        type: integer
      balance:
        description: 'Computed: Total - Allocated'
        type: integer
      created_at:
        $ref: '#/definitions/models.Timestamp'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      total_amount:
        description: 'Computed: Sum of bills/invoices'
        type: integer
      type:
        description: vendor, customer
        type: string
      updated_at:
        $ref: '#/definitions/models.Timestamp'
    type: object
  models.ContactInput:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      type:
        type: string
    type: object
...
  models.Timestamp:
    properties:
      time.Time:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API for managing accounts, contacts, bills, invoices, and transactions.
  title: Portal API
  version: 1.0.0
paths:
...
  /contacts:
    get:
      description: Get a list of all vendors and customers with financial summaries.
      parameters:
      - description: Filter by type (vendor/customer)
        in: query
        name: type
        type: string
      - description: Search by name, email, or phone
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Contact'
                  type: array
              type: object
      security:
      - BasicAuth: []
      summary: List contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      description: Create a new vendor or customer.
      parameters:
      - description: Contact contents
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Contact'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BasicAuth: []
      summary: Create contact
      tags:
      - contacts
  /contacts/{id}:
    delete:
      description: Remove a contact.
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BasicAuth: []
      summary: Delete contact
      tags:
      - contacts
    get:
      description: Get details and financial summary of a specific contact.
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Contact'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BasicAuth: []
      summary: Get contact
      tags:
      - contacts
    put:
      consumes:
      - application/json
      description: Update details of an existing contact.
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated contact contents
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Contact'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BasicAuth: []
      summary: Update contact
      tags:
      - contacts
...
swagger: "2.0"